Move icons in sub-folder.
