+++ pages/signup.js
@@ -5,7 +5,7 @@ import { supabase } from '../utils/initSupabase';
 import { useUser } from '../components/UserContext';
 import Input from '../components/Input';
 import Button from '../components/Button';
-import Logo from '../components/Logo';
+import Logo from '../components/icons/Logo';
 
 const SignUp = () => {
   const [email, setEmail] = useState('');
